requirements.txt
asgiref==3.7.2
Django==5.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.10.0
